requirements.txt
# DIDGERI-BOOM - TikTok AI Platform
# Core Server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
pydantic>=2.10.0

# Video Editing
ffmpeg-python>=0.2.0

# Speech-to-Text (Captions)
faster-whisper>=1.1.0

# HTTP Client (TikTok API)
httpx>=0.28.0

# Scheduling
apscheduler>=3.10.0

# File Watching
watchdog>=6.0.0

# Data & Analytics
aiofiles>=24.1.0
